+++ NEWS.md
@@ -1,6 +1,6 @@
 # aiscm (0.11.2)
 
-* aiscm/ffmpeg-helpers-test.c: fixed Debian sbuild warning "array subscrit is above array bounds"
+* aiscm/ffmpeg-helpers-test.c: fixed Debian sbuild warning "array subscript is above array bounds"
 
 # aiscm (0.11.1)